requirements.txt
fastapi==0.78.0
celery==5.2.6
python-dotenv==0.20.0
typing_extensions==4.2.0
databases==0.5.5
requests==2.27.1
rich==12.4.1
Jinja2==3.1.2
numpy==1.22.3
# todo: breaking update to 4.0.0
filesplit==3.0.2
treelib==1.6.1
uvicorn
python-multipart
email-validator
fastapi-mail
pandas
PyYAML
aiofiles
aio_pika
aiosqlite
gunicorn==20.1.0
###todo pycryptodome-3.10.1 ????
pycryptodome